add batch updated to agent and web server

+++ internal/server/internal/gaugeserver/internal/storage/internal/pg/bootstrap.go
@@ -19,7 +19,7 @@ counter
 (
 	id SERIAL PRIMARY KEY,
  	name TEXT,
- 	value INTEGER
+ 	value BIGINT
 );
 `
